add object mapper to movie model
